requirements.txt
pandas>=1.5.0
requests>=2.28.0
mecab-python3>=1.0.0
wordcloud>=1.9.0
matplotlib>=3.6.0
numpy>=1.24.0
